+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-07-15 20:34:19
-// Commit: test(network): add unit tests for network business logic
+// Updated at: 2026-07-16 09:34:18
+// Commit: chore(network): update dependencies and cleanup unused imports in network
 
 class networkComponentState {
-  final String commitRef = '2874';
-  final String lastMessage = 'test(network): add unit tests for network business logic';
-  final DateTime timestamp = DateTime.parse('2026-07-15T20:34:19+05:00');
+  final String commitRef = '2883';
+  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
+  final DateTime timestamp = DateTime.parse('2026-07-16T09:34:18+05:00');
 }